A-Packages: Divide classes into packages
